requirements.txt
apturl==0.5.2
asn1crypto==1.4.0
awscli==1.16.210
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
boto3==1.9.200
botocore==1.12.200
Brlapi==0.8.2
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.3.9
command-not-found==0.3
cryptography==2.7
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.5.0
distro-info==1.0
Django==2.1.1
django-bootstrap==0.2.4
django-crispy-forms==1.7.2
django-fields==0.3.0
django-livereload-server==0.3.2
django-multiselectfield==0.1.8
django-prometheus==2.1.0
django-redis-sessions==0.6.2
django-request==1.5.4
django-utils==0.0.2
dnspython==1.16.0
docutils==0.14
duplicity==0.8.17
envs==1.3
fasteners==0.14.1
future==0.18.2
gyp==0.1
httplib2==0.18.1
idna==2.8
jeepney==0.6.0
jmespath==0.10.0
keyring==22.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.16.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
monotonic==1.5
netifaces==0.10.9
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.7.2
pexpect==4.8.0
Pillow==8.1.2
prometheus-client==0.11.0
protobuf==3.12.4
pyasn1==0.4.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
Pygments==2.7.1
PyGObject==3.38.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
python-apt===2.2.0-ubuntu0.21.04.1
python-dateutil==2.8.1
python-debian==0.1.39
python-dotenv==0.18.0
pytz==2021.1
pyxdg==0.27
PyYAML==5.1
redis==3.5.3
reportlab==3.5.66
requests==2.22.0
roman==2.0.0
rsa==3.4.2
s3transfer==0.2.1
SecretStorage==3.3.1
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
ssh-import-id==5.11
systemd-python==234
tornado==6.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.7
uWSGI==2.0.18
wadllib==1.3.5
xdg==5
xkit==0.0.0
